// src/app/core/models/all-categories.interface.ts
export interface IAllCategoriesResponse {
  results: number;
  metadata: Metadata;
  data: ICategory[];
}

export interface Metadata {
  currentPage: number;
  numberOfPages: number;
  limit: number;
}

export interface ICategory {
  _id: string;
  name: string;
  slug: string;
  image: string;
  createdAt: string;
  updatedAt: string;
}